added configproperties
