+++ backend/src/main/java/com/htecgroup/skynest/SkyNestApplication.java
@@ -2,52 +2,19 @@ package com.htecgroup.skynest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.htecgroup.skynest.filter.CustomAuthorizationFilter;
-import com.htecgroup.skynest.util.JwtUtils;
-import com.htecgroup.skynest.util.UrlUtil;
 import org.modelmapper.ModelMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.context.annotation.Bean;
-import org.springframework.http.HttpHeaders;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @SpringBootApplication
+@ConfigurationPropertiesScan("com.htecgroup.skynest.properties")
 public class SkyNestApplication {
 
   public static void main(String[] args) {
     SpringApplication.run(SkyNestApplication.class, args);
-  }
-
-  @Bean
-  public WebMvcConfigurer corsConfigurer() {
-    return new WebMvcConfigurer() {
-      @Override
-      public void addCorsMappings(CorsRegistry registry) {
-        String publicIp = UrlUtil.SKYNEST_HOST;
-        String frontendPort = String.valueOf(UrlUtil.FRONTEND_PORT);
-        registry
-            .addMapping("/**")
-            .allowedMethods("*")
-            .allowedOrigins(
-                "http://localhost:3000", String.format("http://%s:%s", publicIp, frontendPort))
-            .allowedOriginPatterns(
-                "http://localhost:3000/*", String.format("http://%s:%s/*", publicIp, frontendPort))
-            .allowedHeaders(
-                JwtUtils.AUTH_HEADER,
-                HttpHeaders.ORIGIN,
-                HttpHeaders.CONTENT_TYPE,
-                HttpHeaders.ACCEPT)
-            .exposedHeaders(
-                JwtUtils.AUTH_HEADER,
-                HttpHeaders.ACCESS_CONTROL_ALLOW_HEADERS,
-                HttpHeaders.ORIGIN,
-                HttpHeaders.ACCEPT,
-                HttpHeaders.ACCESS_CONTROL_REQUEST_METHOD,
-                HttpHeaders.ACCESS_CONTROL_REQUEST_HEADERS);
-      }
-    };
   }
 
   @Bean
