+++ backend/src/main/resources/application.properties
@@ -1,10 +1,10 @@
 spring.profiles.active=@spring.profiles.active@
 #   spring server
 server.port=${BE_PORT}
-#   URLs
-skynest.host=${SKYNEST_HOST}
-backend.port=${BE_PORT}
-frontend.port=${FE_PORT}
+#   server info
+server-info.host=${SKYNEST_HOST}
+server-info.backend-port=${BE_PORT}
+server-info.frontend-port=${FE_PORT}
 #   Logging
 logging.file.path=.log
 logging.logback.rollingpolicy.max-file-size=10MB
